pyproject.toml
[project]
name = "ml-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.3",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "transformers>=4.53.2",
    "wandb>=0.21.0",
    "wordcloud>=1.9.4",
]
